rokeret rundt på tokens så de ligger i token controller
